+++ changelog.md
@@ -7,6 +7,7 @@
 * Add special error code for message not received within poll time
 * Add topic field for messages
 * Make the topic partition list optional for consumer commit
+* Add `store_offset` to consumer
 
 <a name="0.9.1"></a>
 ## 0.9.1 (2017-04-19)